学生单人单服

+++ platform/pc/src/main/server.js
@@ -3,7 +3,7 @@
  * @desc 处理native服务器和游戏服务器的文件
  * @date 2020-02-18 11:42:29 
  * @Last Modified by: 雪糕
- * @Last Modified time: 2020-03-12 16:31:48
+ * @Last Modified time: 2020-03-14 20:37:20
  */
 const http = require('http');
 const url = require('url');
@@ -138,6 +138,10 @@ async function createGameServer() {
 /** 关闭游戏服务器 */
 async function closeGameServer() {
     return new Promise((resolve, reject) => {
+        if (!nativeServer) {
+            return;
+        }
+
         if (!config.gameServerInited) {
             let cmdStr = "taskkill /im game.exe /f";
             util.runCmd(cmdStr, null, `关闭游戏服务器成功`, "关闭游戏服务器错误");
